% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_object_sizes.R
\docType{data}
\name{allobj}
\alias{allobj}
\title{List all objects in environment by size, only where over 100k}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
allobj
}
\description{
\code{wdwd} is a function that...
}
\keyword{datasets}
