% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{compareColumns}
\alias{compareColumns}
\title{\code{compareColumns} compares column types across two df (e.g., in advance of a  merge) #}
\usage{
compareColumns(df1, df2)
}
\description{
\code{compareColumns} compares column types across two df (e.g., in advance of a  merge) #
}
\details{
https://stackoverflow.com/questions/45743991/r-compare-column-types-between-two-dataframes
}
