% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{group_by_sum}
\alias{group_by_sum}
\title{\code{group_by_sum} Quick 'group by' function to look at NA or 0 values for each year}
\usage{
group_by_sum(df, col, group = year, value = NA, name = "n_NA")
}
\description{
\code{group_by_sum} Quick 'group by' function to look at NA or 0 values for each year
}
