% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{group_mean_conf_int}
\alias{group_mean_conf_int}
\title{\code{group_mean_conf_int} Function to calculate confidence intervals for a variable given grouping variables}
\usage{
group_mean_conf_int(df, var, groups = NULL, se_func = se, ...)
}
\description{
\code{group_mean_conf_int} Function to calculate confidence intervals for a variable given grouping variables
}
