% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{grp_n}
\alias{grp_n}
\title{Grouped summaries, listing of unique groups}
\usage{
grp_n(df, groupvar)
}
\description{
\code{grp_n} counts nonmissing observations of each column by group
}
\examples{
 grp_n(mtcars, cyl)
}
