% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{grp_uniq}
\alias{grp_uniq}
\title{\code{grp_uniq} counts nonmissing \emph{distinct} observations of each column by group}
\usage{
grp_uniq(df, groupvar)
}
\description{
\code{grp_uniq} counts nonmissing \emph{distinct} observations of each column by group
}
\examples{
 grp_n(mtcars, cyl)
}
