% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{sidebyside}
\alias{sidebyside}
\title{\code{sidebyside} just lines up 2 or more data frames for comparison view}
\usage{
sidebyside(..., width = 60)
}
\description{
\code{sidebyside} just lines up 2 or more data frames for comparison view
}
\examples{
sidebyside(mtcars[,1:2], mtcars[,3:4])
}
