% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{sm}
\alias{sm}
\title{\code{sm} selects variables matching string 'x', not case-sensitive}
\usage{
sm(df, X)
}
\description{
\code{sm} selects variables matching string 'x', not case-sensitive
}
