% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{smn}
\alias{smn}
\title{\code{smn} creates vector of \emph{names} of variables matching string 'x', not case-sensitive}
\usage{
smn(df, X)
}
\description{
\code{smn} creates vector of \emph{names} of variables matching string 'x', not case-sensitive
}
