% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{Snotm}
\alias{Snotm}
\title{\code{Snotm} selects variables \emph{not} matching that string,  case-sensitive}
\usage{
Snotm(df, X)
}
\description{
\code{Snotm} selects variables \emph{not} matching that string,  case-sensitive
}
