% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_add.R
\name{summ_add}
\alias{summ_add}
\title{Calculate mean, n, sd, and se}
\usage{
summ_add(df, var)
}
\arguments{
\item{var}{}
}
\description{
\code{sum_add} takes a dataframe and a column name, and calculates mean, median, n,  sd, and se
}
\examples{
# Here is an example
x <- sample(x = c(0, 1),  size = 10, replace = TRUE)
se_bin(x)

}
