% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summs.R
\name{summs}
\alias{summs}
\title{Calculate mean, n, sd, and se}
\usage{
summs(df, var, qlow = 0.2, qhigh = 0.8)
}
\description{
\code{sum_add2} takes a column name, and calculates mean, median, n,  sd, and se  and specified (default p20 and p80) quantiles
}
\examples{
# Here is an example
x <- sample(x = c(0, 1),  size = 10, replace = TRUE)
se_bin(x)

}
