% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{tabg}
\alias{tabg}
\title{\code{tabg} --  tabyl one way plus sort by descending frequency -- the version we normally want}
\usage{
tabg(df, col)
}
\description{
\code{tabg} --  tabyl one way plus sort by descending frequency -- the version we normally want
}
