% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{tabsums}
\alias{tabsums}
\title{\code{tabsums} Quick 'group by and summarize'}
\usage{
tabsums(df, yvar, xvar)
}
\description{
\code{tabsums} Quick 'group by and summarize'
}
\examples{
tabsums(mtcars, mpg, cyl)
}
