% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_package.R
\name{XXX}
\alias{XXX}
\title{XXX Explanation in one sentence}
\usage{
XXX(x)
}
\arguments{
\item{XXX}{describe an input to function}
}
\description{
\code{wdwd} is a function that...
}
\details{
XXX This function does nothing. It relies on no packages
}
\examples{
 XXX nothing(mtcars)
}
