% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_function_wip.R
\name{yfind}
\alias{yfind}
\title{A variety of functions for dynamic use
\code{yfind} -- help find where a column with a certain string name takes values across years (or other grouping)}
\usage{
yfind(df = eas_all, text, n = 3, y = year)
}
\description{
A variety of functions for dynamic use
\code{yfind} -- help find where a column with a certain string name takes values across years (or other grouping)
}
